E/middleware hotfix (#224)
Two changes:
(1) Bug fix in middleware, now checks that a supplied locale is in the list of approved locales using determineLocale() rather than just using .includes(). This allows us to check for variants of supported locales (e.g. if you support en then you should be able to handle a user requesting en-US. (Additionally, we now forward source maps if detected, but this is pretty minor)
(2) Throw a build error if your gt.config.json does not match the parameters you have supplied to withGTConfig() which we noticed was a common point of confusion for new users.